requirements.txt
boto3==1.9.204
botocore==1.12.204
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-extra-views==0.12.0
django-storages==1.7.1
djangorestframework==3.10.2
docutils==0.15.2
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
Pillow==6.1.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
